+++ national_rail/national_rail.py
@@ -1,9 +1,9 @@
 """ Module for running the pipeline for extracting incident data from the NationalRail API."""
 import logging
 
-import extract
-import transform
-import load
+import extract_national as extract_national
+import transform_national as transform_national
+import load_national as load_national
 
 FILENAME = "data.xml"
 
@@ -11,8 +11,8 @@ if __name__ == "__main__":
     logging.basicConfig(level=logging.INFO,
                         format="%(asctime)s - %(levelname)s - %(message)s")
 
-    extract.get_national_rail_data(FILENAME)
+    extract_national.get_national_rail_data(FILENAME)
 
-    incidents_data = transform.transform_national_rail_data(FILENAME)
+    incidents_data = transform_national.transform_national_rail_data(FILENAME)
 
-    load.load_incidents(incidents_data)
+    load_national.load_incidents(incidents_data)
